pyproject.toml
[tool.poetry]
name = "speedpay-api-v1"
version = "0.1.0"
description = "Basic api for the speedpay fintech services"
authors = ["Tega Ukavwe <justtega97@gmail.com>"]
readme = "README.md"
packages = [{include = "speedpay_api_v1"}]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework-simplejwt = "^5.2.2"
django-cors-headers = "^3.13.0"
Django = "^4.1.4"
black = "^22.12.0"
pytest = "^7.2.0"
drf-spectacular = "^0.24.2"
django-debug-toolbar = "^3.8.1"
python-dotenv = "^0.21.0"
debugpy = "^1.6.4"
pre-commit = "^2.20.0"
isort = "^5.11.1"
autoflake = "^2.0.0"
mysqlclient = "^2.1.1"
pytest-django = "^4.5.2"
pytest-cov = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
